# app/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Storage Api V1
  description: |-
    Storage Api V1 resources
  version: 1.0.0
paths:
  /health:
    get:
      tags:
        - Health Check
      summary: Check application status
      description: Check application status
      operationId: healthGet
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        '500':
          description: Error request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"

  /file/upload:
    post:
      tags:
        - File
      summary: Upload a single file
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                resource:
                  type: string
                  format: binary
                dirName:
                  type: string
                fileName:
                  type: string
              required:
                - resource
                - dirName
                - fileName
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadFileResponse"
        '500':
          description: Error in request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"

  /file/download:
    get:
      tags:
        - File
      summary: Download a file by name
      operationId: downloadFile
      parameters:
        - name: filePath
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '500':
          description: Error in request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"

  /files:
    get:
      tags:
        - File
      summary: List available files from storage path
      operationId: list
      parameters:
        - name: resource
          in: query
          required: false
          schema:
            type: string
            default: "/"

      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Data'
        '500':
          description: Error in request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"

components:
  schemas:
    Data:
      type: object
      properties:
        files:
          items:
            type: string
      required:
        - files
    UploadFileResponse:
      type: object
      properties:
        fileName:
          type: string
        fileDownloadUri:
          type: string
        contentType:
          type: string
        size:
          type: integer
          format: int64
        extension:
          type: string
    Response:
      type: object
      properties:
        timestamp:
          type: integer
          format: int64
        message:
          type: string
        success:
          type: boolean
        errorInfo:
          type: object
      required:
        - timestamp
        - message
        - success